[ci skip] Incremental commit. Please do not base any external work off of this commit.

+++ cloudfoundry-client/src/main/java/org/cloudfoundry/client/v2/applications/_Resource.java
@@ -54,6 +54,6 @@ abstract class _Resource {
      */
     @JsonProperty("size")
     @Nullable
-    abstract Integer getSize();
+    abstract Long getSize();
 
 }
