+++ cloudfoundry-client/src/main/java/org/cloudfoundry/client/v2/resourcematch/_Resource.java
@@ -48,5 +48,5 @@ abstract class _Resource {
      */
     @JsonProperty("size")
     @Nullable
-    abstract Integer getSize();
+    abstract Long getSize();
 }
